Altering slider variable to be in state namespace
The commander-slider macro was using the namespace `$:/state-reveal-` to
contain the state of whether the slider was expanded or not. However,
this creates a tiddler that core doesn't recognize as a state tiddler.

On NodeJS, this means clicking on that little arrow will write to disk.

On standalone tiddler files, this means clicking the arrow flags your
entire tiddlywiki as having unsaved changes.

By using the $:/state/ namespace for nonpersistent state, the slider's
non-persistent expand state will be treated as non-persistent by
Tiddlywiki.